fix: work on date logic cleanup
